requirements.txt
black==24.3.0
blinker==1.7.0
click==8.1.7
flake8==7.0.0
Flask==3.0.2
gunicorn==21.2.0
install==1.3.5
itsdangerous==2.1.2
jaraco.path==3.7.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.2.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
Werkzeug==3.0.2
